sock: use nonblock connect
With previous implementation connect was blocking, time was not
deterministic and depending on global networking settings.

Checked VM to VM connect blocking time just to have some reference.
1) with allowed traffic; few ms (4-6)
2) not allowed traffic from 4420/tcp; as above; EHOSTUNREACH errno
3) no listener; as above, ECONNREFUSED errno
4) drop rule on target; 135s; ETIMEDOUT errno
5) link down on target; 70s; EHOSTUNREACH errno
6) link down on target + arp table flush; 3s; EHOSTUNREACH errno
7) non existing ip; as above

Blocking behavior is maintained with this change however it is done on
posix poll() and not connect(). Following patch will introduce timeout
opt. 1-3 scenarios results are the same with the change, however 4-7
can be significantly mitigated.

This is a step in the middle of transitioning to async connect.

Change-Id: Ia2f347a3cfdcd427b8302e0220448b0f34d19dd4
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25794
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
(cherry picked from commit 2dff973682aaf7d8a02d3faeba6615aeecb43e00)