sock/posix: return errno instead -1 on ssl conf
This is a prework change for future updates where more precise errors
will be returned to the user.

Note that passing the SSL error directly is not an option, as SSL errors
do not use errno. Instead, EPROTO was chosen as a generic error code for
SSL-related issues.

As part of this change, errno was also removed from log messages, since
the SSL utility functions do not set it.

Change-Id: I7d47eb0ad0d93292e7a98ad4460e0944dba6d2cc
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26105
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
(cherry picked from commit 5dc69b179766db9553656cc07d84906ff83d9bb5)