nvme/tcp: fix lack of qpair flags clear
Issue found by code analysis.

A problematic scenario: after the qpair is enabled, the icreq_send_ack
flag is set. On reset, a reconnect occurs, but the flag is not cleared.
This may lead to unexpected or invalid behavior.

The fix ensures the flag is always cleared during the initialization
path. It is not cleared during disconnect, as we may still need to
respect the in_connect_poll flag.

Change-Id: I2365b02dc7f1752bb515a74b6db45fcff69239e2
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26140
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
(cherry picked from commit 88dd37b1dd18cff15341871bdee203df8091623a)